app/requirements.txt
asgiref==3.5.0
click==8.0.4
Django==3.2.12
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
typed-ast==1.5.2
typing_extensions==4.1.1
zipp==3.7.0
